Give the registry the shared docs shell at its subdomain home
