+++ docs/app/(registry)/registry/page.tsx
