+++ docs/app/(registry)/registry/registry-markdown.tsx
