+++ docs/app/(registry)/registry/shiki-code.tsx
